refactor: Rename firebase to firebaseConfig
