+++ web-client/src/ducks/MyOffers/actions.ts
@@ -1,4 +1,3 @@
-import { firestore } from 'src/firebase';
 import { IPost } from 'src/models/posts/IPost';
 import { Post } from 'src/models/posts/Post';
 import { User } from 'src/models/users';
