+++ web-client/src/ducks/auth/google/functions.ts
@@ -1,4 +1,4 @@
-import firebase, { firebaseAuth } from 'src/firebase';
+import firebase, { firebaseAuth } from 'src/firebaseConfig';
 
 export const googleLoginWithFirebasePopUp = async (): Promise<firebase.auth.UserCredential> => {
   const provider = new firebase.auth.GoogleAuthProvider();
