+++ web-client/src/modules/timeline/containers/TimelineViewContainer/TimelineViewContainer.tsx
@@ -6,7 +6,7 @@ import { MyOffersState } from 'src/ducks/MyOffers/types';
 import { MyRequestsState } from 'src/ducks/MyRequests/types';
 import { ProfileState } from 'src/ducks/profile/types';
 import { getPostWithOffersAndTimelineItems } from 'src/ducks/timeline/functions';
-import { firestore as firestore2 } from 'src/firebase';
+import { firebaseFirestore as firestore2 } from 'src/firebaseConfig';
 import { Post } from 'src/models/posts/Post';
 import { ApplicationPreference } from 'src/models/users';
 import {
