+++ web-client/src/services/search.ts
@@ -1,10 +1,10 @@
 import { IgetSearchKey, IgetSearchKeyResult } from 'src/ducks/search/types';
-import { functions } from 'src/firebase';
+import { firebaseFunctions } from 'src/firebaseConfig';
 
 export const getSearchKey = async ({
   authenticated,
 }: IgetSearchKey): Promise<IgetSearchKeyResult> =>
-  functions
+  firebaseFunctions
     .httpsCallable('https-api-getSearchKey')({
       authenticated,
     })
